+++ examples/text/text_font_spritefont.c
@@ -38,14 +38,14 @@ int main(void)
     Font font2 = LoadFont("resources/custom_alagard.png");        // Font loading
     Font font3 = LoadFont("resources/custom_jupiter_crash.png");  // Font loading
 
-    Vector2 fontPosition1 = { screenWidth/2 - MeasureTextEx(font1, msg1, font1.baseSize, -3).x/2,
+    Vector2 fontPosition1 = { screenWidth/2 - MeasureTextEx(font1, msg1, (float)font1.baseSize, -3).x/2,
                               screenHeight/2 - font1.baseSize/2 - 80 };
 
-    Vector2 fontPosition2 = { screenWidth/2 - MeasureTextEx(font2, msg2, font2.baseSize, -2).x/2,
+    Vector2 fontPosition2 = { screenWidth/2 - MeasureTextEx(font2, msg2, (float)font2.baseSize, -2).x/2,
                               screenHeight/2 - font2.baseSize/2 - 10 };
 
-    Vector2 fontPosition3 = { screenWidth/2 - MeasureTextEx(font3, msg3, font3.baseSize, 2).x/2,
-                              screenHeight/2 - font3.baseSize/2 + 50 };
+    Vector2 fontPosition3 = { screenWidth/2.0f - MeasureTextEx(font3, msg3, (float)font3.baseSize, 2).x/2,
+                              screenHeight/2.0f - font3.baseSize/2 + 50 };
 
     SetTargetFPS(60);               // Set our game to run at 60 frames-per-second
     //--------------------------------------------------------------------------------------
@@ -64,9 +64,9 @@ int main(void)
 
             ClearBackground(RAYWHITE);
 
-            DrawTextEx(font1, msg1, fontPosition1, font1.baseSize, -3, WHITE);
-            DrawTextEx(font2, msg2, fontPosition2, font2.baseSize, -2, WHITE);
-            DrawTextEx(font3, msg3, fontPosition3, font3.baseSize, 2, WHITE);
+            DrawTextEx(font1, msg1, fontPosition1, (float)font1.baseSize, -3, WHITE);
+            DrawTextEx(font2, msg2, fontPosition2, (float)font2.baseSize, -2, WHITE);
+            DrawTextEx(font3, msg3, fontPosition3, (float)font3.baseSize, 2, WHITE);
 
         EndDrawing();
         //----------------------------------------------------------------------------------
